tg_farm_bot/requirements.txt
aiogram==3.5.0
logging==0.4.9.6
python-dotenv==1.0.1